Using abridged game context msg to save bandwidth
Modified how message if constructed on server side to to reduce reduant information. The scheme uses \'checkpoints\' which are messages that contain the full game context (i.e every entity in the scene). Checkpoints are sent once a second. On the client sise, every non-checkpoint game message relies on the last sent checkpoint to reconstruct the full game context